cookie and notify are not defined
When trying to remove a cookie I noticed that the cookie wasn't being removed. Upon further inspection I saw that the function was relying on the local "cookie" variable which wasn't copied across from the initialization code. Same seems to be true for "notify".

Cookies can now be removed happily.