Cleanuped code for enforced cookie usage
In my last pull request I added the first support for handling of the enforced usage of cookies as storage. Now I added this support to the remove function and to the clearAll function. Also I exluded warnings that the browser does not support local storage if the user explicitely wants to use cookies.